+++ web-camera-shop-springmvc-project/src/java/sv/iuh/project/controller/UserController.java
@@ -33,46 +33,66 @@ import sv.iuh.project.service.UserService;
 @Controller
 @RequestMapping("/user")
 public class UserController {
-    
-      @Autowired
-      private UserService userService;
-       @Autowired
-      private LoginService loginService;
-      @RequestMapping("/show")
+
+    @Autowired
+    private UserService userService;
+    @Autowired
+    private LoginService loginService;
+
+    @RequestMapping("/show")
     public String viewUser(ModelMap mm, HttpSession session) {
-        //mm.put("list", userService.getListNav(0, 6));
-        mm.put("list",userService.getListNav(0, 6));
-        return "admin/usermanage";
+        UserShop userShop = (UserShop) session.getAttribute("userlogin");
+        if (userShop != null) {
+            if (userShop.getRole().equals("user")) {
+                return "redirect:/";
+            }
+            if (userShop.getRole().equals("admin")) {
+                //mm.put("list", userService.getListNav(0, 6));
+                mm.put("list", userService.getListNav(0, 6));
+                return "admin/usermanage";
+            } else {
+                return "redirect:/";
+            }
+        } else {
+            return "redirect:/";
+        }
 
     }
-    
+
     @RequestMapping(value = "/show/{page}", method = RequestMethod.GET)
     public String viewProductAdminByPage(ModelMap mm, HttpSession session, @PathVariable("page") int page) {
-        mm.put("list", userService.getListNav((page - 1) * 6, 6));
-        
-        return "admin/usermanage";
+        UserShop userShop = (UserShop) session.getAttribute("userlogin");
+        if (userShop != null) {
+            if (userShop.getRole().equals("user")) {
+                return "redirect:/";
+            }
+            if (userShop.getRole().equals("admin")) {
+                mm.put("list", userService.getListNav((page - 1) * 6, 6));
+                return "admin/usermanage";
+            } else {
+                return "redirect:/";
+            }
+        } else {
+            return "redirect:/";
+        }
+
     }
-    
-    
-    
+
     // thêm khách hàng
-    
-     //Luu san pham
     @RequestMapping(value = "save", method = RequestMethod.POST, produces = "text/plain;charset=UTF-8")
     public String viewProductSave(ModelMap mm, HttpSession session, HttpServletRequest request,
             @RequestParam(value = "image") MultipartFile photo) throws UnsupportedEncodingException {
         request.setCharacterEncoding("UTF-8");
-        
+
         String fullName = request.getParameter("fullname");
         String userName = request.getParameter("userName");
         String passWord = request.getParameter("password");
         String email = request.getParameter("email");
         String phone = request.getParameter("phone");
-       //Date birth =  java.sql.Date.valueOf(request.getParameter("birthday"));
-       
+        Date birth = java.sql.Date.valueOf(request.getParameter("birthday"));
+
         String address = request.getParameter("address");
-        
-        
+
         UserShop userShop = new UserShop(email, passWord, userName, address, fullName, saveFile(photo), phone);
         userShop.setRole("user");
         userShop.setActive(0);
@@ -90,25 +110,24 @@ public class UserController {
             }
             userShop.setUserID(id);
             userService.update(userShop);
-            
+
             String inf = "Sua thanh cong";
             return "redirect:/user/showformupdate?id=" + id + "&success=" + inf + "";
         }
     }
-    
-    
-    
-     //Luu san pham
+
+    //Luu san pham
     @RequestMapping(value = "saveUser", method = RequestMethod.POST, produces = "text/plain;charset=UTF-8")
     public String viewUserSave(ModelMap mm, HttpSession session, HttpServletRequest request) throws UnsupportedEncodingException {
+
         request.setCharacterEncoding("UTF-8");
-        int id=Integer.parseInt(request.getParameter("id"));
+        int id = Integer.parseInt(request.getParameter("id"));
         String fullName = request.getParameter("tendaydu");
         String name = request.getParameter("tendangnhap");
         String pass = request.getParameter("pass");
         String email = request.getParameter("email");
         String phone = request.getParameter("sodt");
-        java.sql.Date birth =  java.sql.Date.valueOf(request.getParameter("ngaysinh"));
+        java.sql.Date birth = java.sql.Date.valueOf(request.getParameter("ngaysinh"));
         String diachi = request.getParameter("diachi");
         String img = request.getParameter("img");
         String address = diachi;
@@ -132,29 +151,37 @@ public class UserController {
             return "user/login";
         }
         session.setAttribute("userlogin", user);
-        mm.addAttribute("user",session.getAttribute("userlogin"));
+        mm.addAttribute("user", session.getAttribute("userlogin"));
         return "redirect:/user/showformuserupdate";
     }
-    
-    
-  
-    
-    
+
     @RequestMapping(value = "/search", method = RequestMethod.GET)
     public String viewSearch(ModelMap mm, HttpSession session, @RequestParam(required = true) String name) {
-        if (name == "" ) {
-            return "redirect:/user/show";
-        } else {
-            mm.put("list", userService.getListByName(name));
+        UserShop userShop = (UserShop) session.getAttribute("userlogin");
+        if (userShop != null) {
+            if (userShop.getRole().equals("user")) {
+                return "redirect:/";
+            }
+            if (userShop.getRole().equals("admin")) {
+                if (name == "") {
+                    return "redirect:/user/show";
+                } else {
+                    mm.put("list", userService.getListByName(name));
 //            mm.put("listbrand", productBrandService.getAll());
 //            mm.put("brand", productBrandService.findById(id));
-            mm.put("s", name);
-            return "admin/usermanage";
+                    mm.put("s", name);
+                    return "admin/usermanage";
+                }
+            } else {
+                return "redirect:/";
+            }
+        } else {
+            return "redirect:/";
         }
+
     }
-    
+
     //hiển thị form
-    
     @RequestMapping(value = "/showuserform", method = RequestMethod.GET)
     public String viewUserUpdateNew(ModelMap mm, HttpSession session) {
         mm.put("list", userService.getAll());
@@ -162,9 +189,7 @@ public class UserController {
         mm.put("r", required);
         return "user/useredit";
     }
-    
-    
-    
+
     @RequestMapping(value = "/showform", method = RequestMethod.GET)
     public String viewUserNew(ModelMap mm, HttpSession session) {
         mm.put("list", userService.getAll());
@@ -172,36 +197,33 @@ public class UserController {
         mm.put("r", required);
         return "admin/userform";
     }
-    
-    
+
     @RequestMapping(value = "/showformupdate")
     public String showFormUdate(ModelMap mm, @RequestParam("id") int id, @RequestParam(required = false) String success) {
         UserShop userShop = userService.findById(id);
-      
-        SimpleDateFormat sm= new SimpleDateFormat("yyy-MM-dd");
-        String date =sm.format(userShop.getBirthday());
-        System.out.println("-------------------------"+date);
+
+        SimpleDateFormat sm = new SimpleDateFormat("yyy-MM-dd");
+        String date = sm.format(userShop.getBirthday());
+        System.out.println("-------------------------" + date);
         mm.put("userShop", userShop);
-        mm.put("date",date);
+        mm.put("date", date);
         mm.put("success", success);
         return "admin/userform";
     }
-    
-    
+
     @RequestMapping(value = "/showformuserupdate")
-    public String showFormUserUdate(ModelMap mm ,HttpSession session) {
-        
-        UserShop userShop=(UserShop) session.getAttribute("userlogin");
-         //int id = Integer.parseInt(request.getParameter("id"));
+    public String showFormUserUdate(ModelMap mm, HttpSession session) {
+
+        UserShop userShop = (UserShop) session.getAttribute("userlogin");
+        //int id = Integer.parseInt(request.getParameter("id"));
         //UserShop userShop = userService.findById(id);
-      
-        
+
         mm.put("userShop", userShop);
-      
-         return "user/useredit";
+
+        return "user/useredit";
     }
-    
-     private String saveFile(MultipartFile file) {
+
+    private String saveFile(MultipartFile file) {
         if (null != file && !file.isEmpty()) {
             try {
                 byte[] bytes = file.getBytes();
@@ -226,19 +248,28 @@ public class UserController {
         }
         return null;
     }
-     
-     @RequestMapping(value = "remove")
+
+    @RequestMapping(value = "remove")
     public String viewRemove(ModelMap mm, HttpSession session, @RequestParam("id") int id) {
-        UserShop p = userService.findById(id);
-        if (p != null) {
-            userService.delete(p);
+        UserShop userShop = (UserShop) session.getAttribute("userlogin");
+        if (userShop != null) {
+            if (userShop.getRole().equals("user")) {
+                return "redirect:/";
+            }
+            if (userShop.getRole().equals("admin")) {
+                UserShop p = userService.findById(id);
+                if (p != null) {
+                    userService.delete(p);
+                }
+                mm.put("list", userService.getAll());
+                return "admin/usermanage";
+            } else {
+                return "redirect:/";
+            }
+        } else {
+            return "redirect:/";
         }
-        mm.put("list", userService.getAll());
-        return "admin/usermanage";
+
     }
-    
-    
-    
-    
-     
+
 }
